MainApplication.kt | AppModule.kt is added || Hilt Constructor Injection is used.
