+++ app/src/main/java/com/example/dagger_hilt_android/model/Battery.kt
@@ -1,8 +1,16 @@
 package com.example.dagger_hilt_android.model
 
+import com.example.dagger_hilt_android.di.MName
 import com.example.dagger_hilt_android.utils.TAG
+import javax.inject.Inject
 
-class Battery(private val name: String) {
+/**
+ * This is an example of Hilt Constructor Injection
+ *
+ * Warning: As the argument of [Battery]
+ * @param name is used in Hilt System you cannot use private modifier
+ * */
+class Battery @Inject constructor(@MName val name: String) {
     fun printBatteryName() {
         println("$TAG :: Battery :: Battery Name: $name")
     }
